[Refactor] CustomView 이용해서 View와 ViewController 파일 분리
